# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Online Coding Interview Platform API
  description: |
    REST API for managing collaborative coding interview sessions.
    Real-time code synchronization is handled via Socket.io WebSocket connections.
  version: 1.0.0
  contact:
    name: API Support
    email: support@codinginterview.com
servers:
  - url: http://localhost:3000
    description: Development server
tags:
  - name: Health
    description: Health check endpoints
  - name: Rooms
    description: Interview room management

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Returns the health status of the API
      operationId: getHealth
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  timestamp:
                    type: string
                    format: date-time
                    example: '2025-12-09T10:30:00.000Z'

  /api/rooms:
    post:
      tags:
        - Rooms
      summary: Create a new interview room
      description: Generates a unique room ID for a new coding interview session
      operationId: createRoom
      responses:
        '201':
          description: Room created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  roomId:
                    type: string
                    format: uuid
                    example: '550e8400-e29b-41d4-a716-446655440000'
                  url:
                    type: string
                    example: '/room/550e8400-e29b-41d4-a716-446655440000'

  /api/rooms/{roomId}:
    get:
      tags:
        - Rooms
      summary: Get room information
      description: Retrieve details about a specific interview room
      operationId: getRoomInfo
      parameters:
        - name: roomId
          in: path
          required: true
          description: Unique identifier of the room
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Room information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  roomId:
                    type: string
                    format: uuid
                    example: '550e8400-e29b-41d4-a716-446655440000'
                  participantCount:
                    type: integer
                    example: 2
                  createdAt:
                    type: string
                    format: date-time
                    example: '2025-12-09T10:30:00.000Z'
        '404':
          description: Room not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Room not found'

components:
  schemas:
    Room:
      type: object
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        participants:
          type: array
          items:
            $ref: '#/components/schemas/Participant'
        currentCode:
          type: string
    
    Participant:
      type: object
      properties:
        socketId:
          type: string
        joinedAt:
          type: string
          format: date-time

  securitySchemes: {}

# WebSocket Events Documentation (not part of OpenAPI standard, but documented here)
x-websocket-events:
  client-to-server:
    join-room:
      description: Client joins a specific room
      payload:
        type: string
        description: Room ID (UUID)
    code-change:
      description: Client sends code changes to sync with other participants
      payload:
        type: object
        properties:
          roomId:
            type: string
          code:
            type: string
    cursor-change:
      description: Client sends cursor position for collaborative editing indicators
      payload:
        type: object
        properties:
          roomId:
            type: string
          position:
            type: object
    code-executed:
      description: Notify that code was executed (for analytics)
      payload:
        type: object
        properties:
          roomId:
            type: string
          timestamp:
            type: string
  
  server-to-client:
    load-code:
      description: Server sends current code state when client joins
      payload:
        type: object
        properties:
          code:
            type: string
    code-update:
      description: Server broadcasts code changes to other participants
      payload:
        type: object
        properties:
          code:
            type: string
    user-joined:
      description: Notify that a new user joined the room
      payload:
        type: object
        properties:
          participantCount:
            type: integer
    user-left:
      description: Notify that a user left the room
      payload:
        type: object
        properties:
          participantCount:
            type: integer
    room-info:
      description: Room metadata sent to client
      payload:
        type: object
        properties:
          participantCount:
            type: integer
    room-error:
      description: Error message related to room operations
      payload:
        type: object
        properties:
          message:
            type: string